requirements.txt
absl-py==2.1.0
accelerate==1.1.0
addict==2.4.0
aggdraw==1.3.19
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.46.1
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==4.0.0
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.2.2
executing==2.1.0
facexlib==0.3.0
filelock==3.16.0
filterpy==1.4.5
flatbuffers==25.1.24
fonttools==4.53.1
frozenlist==1.7.0
fsspec==2024.9.0
ftfy==6.3.1
future==1.0.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
hf-xet==1.1.5
huggingface-hub==0.33.4
icecream==2.1.5
identify==2.6.12
idna==3.10
imageio==2.35.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lmdb==1.7.2
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.9.1
numba==0.61.2
numpy==2.0.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai-clip==1.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.14.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pi==0.1.2
pillow==10.4.0
piq==0.8.0
platformdirs==4.3.2
pluggy==1.6.0
POT==0.9.5
pre_commit==4.2.0
prompt_toolkit==3.0.47
propcache==0.3.2
protobuf==5.28.1
psutil==6.0.0
ptflops==0.7.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
Pygments==2.18.0
pyiqa==0.1.14.1
pyparsing==3.1.4
pytest==8.4.1
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.12.3
safetensors==0.5.3
scikit-dimension==0.3.4
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.14.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.17
tokenizers==0.15.2
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.37.2
triton==3.3.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
virtualenv==20.31.2
visualkeras==0.1.4
wandb==0.18.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
